(Feat) Make common trading runner to be used by backtests and live trading
