+++ trading/algorithms/jenetic_segmentation_oscillatory_heuristics.py
@@ -1,12 +1,10 @@
-import datetime
 import math
 from decimal import Decimal
 
 from bson import ObjectId
 
 from trading.algorithms.base import Strategy
-from trading.broker import MarketOrder, ORDER_MARKET, SIDE_BUY, SIDE_SELL, SIDE_STAY, PRICE_ASK_CLOSE, PRICE_ASK_HIGH, \
-    PRICE_ASK_LOW, PRICE_ASK
+from trading.broker import SIDE_BUY, SIDE_SELL, SIDE_STAY, PRICE_ASK_CLOSE, PRICE_ASK_HIGH, PRICE_ASK_LOW, PRICE_ASK
 from trading.broker.constants import GRANULARITY_TEN_MINUTE
 from trading.indicators import calc_chandalier_exits, INTERVAL_FORTY_CANDLES
 from trading.indicators.overlap_studies import calc_moving_average
@@ -109,23 +107,6 @@ class Josh(Strategy):
             order = self.make_order(asking_price, order_side=order_decision)
 
         return order_decision, order
-
-    def make_order(self, asking_price, order_side=SIDE_BUY):
-        trade_expire = datetime.datetime.utcnow() + datetime.timedelta(days=1)
-        trade_expire = trade_expire.isoformat('T') + 'Z'
-
-        if order_side == SIDE_BUY:
-            units = self.calc_units_to_buy(asking_price)
-        else:
-            units = self.calc_units_to_sell(asking_price)
-
-        instrument = self.portfolio.instrument
-        side = order_side
-        order_type = ORDER_MARKET
-        price = asking_price
-        expiry = trade_expire
-
-        return MarketOrder(instrument, units, side, order_type, price, expiry)
 
     def shutdown(self, started_at, ended_at, num_ticks, num_orders, shutdown_cause):
         session_info = self.make_trading_session_info(started_at, ended_at, num_ticks, num_orders, shutdown_cause)
