+++ trading/algorithms/moving_average_crossover.py
@@ -1,11 +1,10 @@
-import datetime
 import math
 from decimal import Decimal
 
 from bson import ObjectId
 
 from trading.algorithms.base import Strategy
-from trading.broker import MarketOrder, ORDER_MARKET, SIDE_BUY, SIDE_SELL, SIDE_STAY, PRICE_ASK_CLOSE, PRICE_ASK
+from trading.broker import SIDE_BUY, SIDE_SELL, SIDE_STAY, PRICE_ASK_CLOSE, PRICE_ASK
 from trading.indicators import INTERVAL_TEN_CANDLES, INTERVAL_TWENTY_CANDLES
 from trading.indicators.overlap_studies import calc_moving_average
 from trading.util.transformations import normalize_price_data, normalize_current_price_data
@@ -56,9 +55,6 @@ class MAC(Strategy):
         short_ma = Decimal(calc_moving_average(closing_market_data, INTERVAL_TEN_CANDLES))
         long_ma = Decimal(calc_moving_average(closing_market_data, INTERVAL_TWENTY_CANDLES))
 
-        if math.isnan(long_ma):
-            self.logger.error('JJDEBUG: Closing Market Data', data=closing_market_data)
-
         self.strategy_data['asking_price'] = asking_price
         self.strategy_data['short_term_ma'] = short_ma
         self.strategy_data['long_term_ma'] = long_ma
@@ -75,7 +71,6 @@ class MAC(Strategy):
 
         try:
             diff = 100 * (short_term - long_term) / ((short_term + long_term) / 2)
-            self.logger.info('Diff {diff}'.format(diff=diff))
 
             if diff >= self.crossover_threshold and not self.invested:
                 decision = SIDE_BUY
@@ -96,25 +91,6 @@ class MAC(Strategy):
             order = None
 
         return decision, order
-
-    def make_order(self, asking_price, order_side=SIDE_BUY):
-        trade_expire = datetime.datetime.utcnow() + datetime.timedelta(days=1)
-        trade_expire = trade_expire.isoformat("T") + "Z"
-
-        if order_side == SIDE_BUY:
-            units = self.calc_units_to_buy(asking_price)
-        else:
-            units = self.calc_units_to_sell(asking_price)
-
-        self.logger.info('Calculated units {units} and side {side}'.format(units=units, side=order_side))
-
-        instrument = self.portfolio.instrument
-        side = order_side
-        order_type = ORDER_MARKET
-        price = asking_price
-        expiry = trade_expire
-
-        return MarketOrder(instrument, units, side, order_type, price, expiry)
 
     def shutdown(self, started_at, ended_at, num_ticks, num_orders, shutdown_cause):
         session_info = self.make_trading_session_info(started_at, ended_at, num_ticks, num_orders, shutdown_cause)
