+++ trading/live/exceptions.py
